// src/app/backend/model/userIdentityImageType.ts
/**
 * OutOut API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type UserIdentityImageType = 0 | 1;

export const UserIdentityImageType = {
    Brand: 0 as UserIdentityImageType,
    Contact: 1 as UserIdentityImageType
};